Synchronize the Chrony server's NTS certificate with client machines
Also, clearly distinguish between Chrony server and client for NTS
setup.

Note: this is a temporal configuration that depends on the Chrony server
being setup prior to the client. This may need to be refactored in the
future to remove this temporal dependency

Solution: this could be refactored to upload certificates from the
Chrony server to a central repository, and then instruct the client to
fetch the certificates from the central repository with a retry
mechanism. Alternatively, a configuration management tool or service
discovery mechanism could be implemented that dynamically updates client
configurations with the necessary server information and certificates
(e.g.: Consul by HashCorp or etcd). This would require this role to
depend on another role (e.g.: ansible-role-consul or ansible-role-etcd)